feat: 统一示例
