+++ playground/src/pages/examples/basic/index.vue
@@ -5,13 +5,23 @@
 </template>
 
 <script lang="ts" setup>
+import type { PieSeriesOption } from "echarts/charts";
 import { PieChart } from "echarts/charts";
+import type { DatasetComponentOption, LegendComponentOption, TooltipComponentOption } from "echarts/components";
 import { DatasetComponent, LegendComponent, TooltipComponent } from "echarts/components";
-import * as echarts from "echarts/core";
+import type { ComposeOption } from "echarts/core";
+import { use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
-import { getData } from "./data.ts";
+import { GLOBAL_OPTION } from "../echarts.ts";
 
-echarts.use([
+type EChartsOption = ComposeOption<
+  | LegendComponentOption
+  | TooltipComponentOption
+  | DatasetComponentOption
+  | PieSeriesOption
+>;
+
+use([
   LegendComponent,
   TooltipComponent,
   DatasetComponent,
@@ -19,7 +29,52 @@ echarts.use([
   CanvasRenderer
 ]);
 
-const option = shallowRef(getData());
+const option = ref({
+  ...GLOBAL_OPTION,
+  legend: {
+    top: 10,
+    left: "center"
+  },
+  tooltip: {
+    trigger: "item",
+    textStyle: {
+      // #ifdef MP-WEIXIN
+      textShadowBlur: 1
+      // #endif
+    }
+  },
+  series: [
+    {
+      type: "pie",
+      radius: ["30%", "52%"],
+      label: {
+        show: false,
+        position: "center"
+      },
+      itemStyle: {
+        borderWidth: 2,
+        borderColor: "#ffffff",
+        borderRadius: 10
+      },
+      emphasis: {
+        label: {
+          show: true,
+          fontSize: 20
+        }
+      }
+    }
+  ],
+  dataset: {
+    dimensions: ["来源", "数量"],
+    source: [
+      ["Search Engine", 1048],
+      ["Direct", 735],
+      ["Email", 580],
+      ["Union Ads", 484],
+      ["Video Ads", 300]
+    ]
+  }
+} satisfies EChartsOption);
 </script>
 
 <route>
