+++ playground/src/pages/examples/builtin-theme/index.vue
@@ -25,21 +25,30 @@
 </template>
 
 <script lang="ts" setup>
+import type { LineSeriesOption } from "echarts/charts";
 import { LineChart } from "echarts/charts";
-import {
-  DatasetComponent,
-  LegendComponent,
-  PolarComponent,
-  TitleComponent,
-  TooltipComponent
+import type {
+  DatasetComponentOption,
+  LegendComponentOption,
+  PolarComponentOption,
+  TooltipComponentOption
 } from "echarts/components";
-import * as echarts from "echarts/core";
+import { DatasetComponent, LegendComponent, PolarComponent, TooltipComponent } from "echarts/components";
+import type { ComposeOption } from "echarts/core";
+import { use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
 import type { CSSProperties } from "vue";
-import { getData } from "./data.ts";
+import { GLOBAL_OPTION } from "../echarts.ts";
 
-echarts.use([
-  TitleComponent,
+type EChartsOption = ComposeOption<
+  | LegendComponentOption
+  | TooltipComponentOption
+  | PolarComponentOption
+  | DatasetComponentOption
+  | LineSeriesOption
+>;
+
+use([
   LegendComponent,
   TooltipComponent,
   PolarComponent,
@@ -48,22 +57,65 @@ echarts.use([
   CanvasRenderer
 ]);
 
-const option = shallowRef(getData());
+const option = ref({
+  ...GLOBAL_OPTION,
+  legend: {
+    top: 10,
+    left: "center"
+  },
+  tooltip: {
+    trigger: "axis",
+    axisPointer: {
+      type: "cross"
+    },
+    textStyle: {
+      // #ifdef MP-WEIXIN
+      textShadowBlur: 1
+      // #endif
+    }
+  },
+  polar: {
+    radius: "60%",
+    center: ["50%", "56%"]
+  },
+  angleAxis: {
+    type: "value",
+    startAngle: 0
+  },
+  radiusAxis: {
+    min: 0
+  },
+  series: [
+    {
+      type: "line",
+      coordinateSystem: "polar",
+      showSymbol: false
+    }
+  ],
+  dataset: {
+    dimensions: ["数值"],
+    source: getData()
+  }
+} satisfies EChartsOption);
 
 const theme = shallowRef("dark");
 
 const loading = shallowRef(false);
 
 const loadingOptions = computed(() => {
-  if (theme.value !== "dark") {
-    return undefined;
+  const value = {
+    text: "请稍候..."
+  };
+
+  if (theme.value === "dark") {
+    Object.assign(value, {
+      color: "#ffffff",
+      textColor: "#ffffff",
+      maskColor: "rgba(0, 0, 0, 0.7)"
+    });
   }
 
-  return {
-    color: "#ffffff",
-    textColor: "#ffffff",
-    maskColor: "rgba(0, 0, 0, 0.7)"
-  };
+  return value;
 });
 
 const styles = computed(() => {
@@ -79,6 +131,19 @@ const styles = computed(() => {
 
   return value;
 });
+
+function getData() {
+  const data: [number, number][] = [];
+
+  for (let i = 0; i <= 360; i += 1) {
+    const t = (i / 180) * Math.PI;
+    const r = Math.sin(2 * t) * Math.cos(2 * t);
+
+    data.push([r, i]);
+  }
+
+  return data;
+}
 </script>
 
 <route>
