+++ playground/src/pages/examples/echarts.ts
@@ -1,6 +1,5 @@
 export const GLOBAL_OPTION = {
   textStyle: {
-    fontFamily: `MiSans, Inter, "Helvetica Neue", Helvetica, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "微软雅黑", Arial, sans-serif`,
-    fontWeight: 300
+    fontFamily: `MiSans, Inter, "Helvetica Neue", Helvetica, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "微软雅黑", Arial, sans-serif`
   }
 };
