+++ playground/src/pages/examples/export-image/index.vue
@@ -17,12 +17,17 @@
 </template>
 
 <script lang="ts" setup>
+import type { LineSeriesOption } from "echarts/charts";
 import { LineChart } from "echarts/charts";
+import type { PolarComponentOption } from "echarts/components";
 import { PolarComponent } from "echarts/components";
+import type { ComposeOption } from "echarts/core";
 import { use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
-import { getData } from "./data.ts";
+import { GLOBAL_OPTION } from "../echarts.ts";
 import type { UniEchartsInst } from "@/uni_modules/xiaohe-echarts";
+
+type EChartsOption = ComposeOption<PolarComponentOption | LineSeriesOption>;
 
 use([
   PolarComponent,
@@ -32,7 +37,36 @@ use([
 
 const chartEl = ref<UniEchartsInst | null>(null);
 
-const option = shallowRef(getData());
+const option = ref({
+  ...GLOBAL_OPTION,
+  backgroundColor: "#ffffff",
+  polar: {},
+  angleAxis: {
+    type: "value",
+    startAngle: 0
+  },
+  radiusAxis: {},
+  series: [
+    {
+      type: "line",
+      coordinateSystem: "polar",
+      data: getData()
+    }
+  ]
+} satisfies EChartsOption);
+
+function getData() {
+  const data: [number, number][] = [];
+
+  for (let i = 0; i <= 100; i += 1) {
+    const theta = (i / 100) * 360;
+    const r = 5 * (1 + Math.sin((theta / 180) * Math.PI));
+
+    data.push([r, theta]);
+  }
+
+  return data;
+}
 
 async function saveAsImage() {
   if (chartEl.value == null) {
